Refactor the Request locale API
Following the addition of:
 - `List<Locale> locales(BiFunction<List<Locale.LanguageRange>, List<Locale>, List<Locale>> filter`
 - `Locale locale(BiFunction<List<Locale.LanguageRange>, List<Locale>, Locale> filter)`

`Locale locale(Iterable<Locale> locales)` and `Locale locale(Locale... locales)`
were made redundant by the new two functions. This use case can be implemented
in the following way:

    List<Locale> supportedLocales = ...
    Locale locale = req.locale((ranges, locales) ->
      Locale.lookup(ranges, supportedLocales));

`List<Locale> locales()` is just an shorthand for `locales(Locale::filter)`.

`Locale locale()` simply returns the first element of `locales` or the default
locale as per `app.lang` if there are no matches. This function is expected to
be the most used one.

Issue #273