Change default behavior of Request.locale()
Now uses `Locale#lookup` instead of `Locale#filter` + findFirst and
fallback. As a result, it will return the locale with the closest match
in case that is possible.

Example: with a language setting of "en, de", a request with the
`Accept-Language` header set to `de-CH` will now return `de` instead of
`en`.